+++ core/src/main/java/org/realityforge/arez/spy/ObservableChangedEvent.java
@@ -3,6 +3,8 @@ package org.realityforge.arez.spy;
 import java.util.Map;
 import java.util.Objects;
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import org.realityforge.arez.Arez;
 import org.realityforge.arez.Observable;
 
 /**
@@ -15,16 +17,25 @@ public final class ObservableChangedEvent
 
   @Nonnull
   private final Observable<?> _observable;
+  @Nullable
+  private final Object _value;
 
-  public ObservableChangedEvent( @Nonnull final Observable<?> observable )
+  public ObservableChangedEvent( @Nonnull final Observable<?> observable, @Nullable final Object value )
   {
     _observable = Objects.requireNonNull( observable );
+    _value = value;
   }
 
   @Nonnull
   public Observable<?> getObservable()
   {
     return _observable;
+  }
+
+  @Nullable
+  public Object getValue()
+  {
+    return _value;
   }
 
   /**
@@ -35,5 +46,9 @@ public final class ObservableChangedEvent
   {
     map.put( "type", TYPE_NAME );
     map.put( "observable", getObservable().getName() );
+    if ( Arez.arePropertyIntrospectorsEnabled() )
+    {
+      map.put( "value", getValue() );
+    }
   }
 }